+++ pkg/koyeb/instances_logs.go
@@ -2,12 +2,13 @@ package koyeb
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/errors"
 	"github.com/spf13/cobra"
 )
 
-func (h *InstanceHandler) Logs(ctx *CLIContext, cmd *cobra.Command, args []string) error {
+func (h *InstanceHandler) Logs(ctx *CLIContext, cmd *cobra.Command, since time.Time, args []string) error {
 	instance, err := h.ResolveInstanceArgs(ctx, args[0])
 	if err != nil {
 		return err
@@ -26,6 +27,7 @@ func (h *InstanceHandler) Logs(ctx *CLIContext, cmd *cobra.Command, args []strin
 		"",
 		"",
 		instanceDetail.Instance.GetId(),
+		since,
 		GetBoolFlags(cmd, "full"),
 	)
 	if err != nil {